backend/requirements.txt
pandas
dspy
accelerate
sentence-transformers
weaviate-client
rich
psutil
tqdm
fastapi
uvicorn
